break-up analyser component into a few smaller ones, and refactor look of progress collapsible
