+++ web-ui/src/components/analyser/analyser.less
@@ -27,13 +27,6 @@
   height: 50px;
 }
 
-.tag {
-  position: relative;
-  left: 0px;
-  top: 6px;
-  z-index: 2;
-}
-
 .title {
   font-size: 30px;
   margin-top: 0;
