+++ web-ui/src/components/analyser/lecture-progress.less
@@ -1,0 +1,19 @@
+
+.tag {
+  position: relative;
+  left: 0px;
+  top: 6px;
+  z-index: 2;
+}
+
+.collapsible {
+  width: 100%;
+
+  > div > div > span {
+    width: 90% !important;
+  }
+}
+
+.overall {
+  width: 100%;
+}
