+++ web-ui/src/components/preview.less
@@ -7,7 +7,20 @@
   }
 }
 
+.stat * h3 {
+  margin: 0;
+  padding: 0;
+}
+
 .language {
+  color: black;
+}
+
+.length {
+  color: black;
+}
+
+.words {
   color: black;
 }
